refactor: reorganize calendar utility imports and create new calendar and calendar-drag modules
